+++ internal/domain/message.go
@@ -3,9 +3,9 @@ package domain
 import "time"
 
 type Message struct {
-	ID        int
-	RoomID    int
-	UserID    int
+	ID        int64
+	RoomID    int64
+	UserID    int64
 	Text      string
 	CreatedAt time.Time
 }
